+++ src/main/java/com/example/demo/mapper/SupplyMapper.java
@@ -1,20 +1,31 @@
 package com.example.demo.mapper;
 
 import com.example.demo.model.dto.response.SupplyResponse;
-import com.example.demo.model.entity.*;
-import com.example.demo.repository.ProductRepository;
+import com.example.demo.model.entity.CustomerEntity;
+import com.example.demo.model.entity.DealEntity;
+import com.example.demo.model.entity.ProductEntity;
+import com.example.demo.model.entity.SellerEntity;
+import com.example.demo.model.entity.SupplyEntity;
 import jakarta.transaction.Transactional;
-import org.mapstruct.Context;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 @Mapper(componentModel = "spring")
 public abstract class SupplyMapper {
 
-    public abstract SupplyResponse mapToSupplyResponse(SupplyEntity supplyEntity);
+    @Mapping(target = "name", source = "productEntity.name")
+    public SupplyResponse mapToSupplyResponse(SupplyEntity supplyEntity) {
+        return SupplyResponse.builder()
+                .weight(supplyEntity.getWeight())
+                .pricePaid(supplyEntity.getPricePaid())
+                .name(supplyEntity.getProductEntity().getName())
+                .cost(supplyEntity.getPricePaid().multiply(BigDecimal.valueOf(supplyEntity.getWeight())))
+                .build();
+    }
 
     @Transactional
     public SupplyEntity mapToSupplyEntity(SellerEntity sellerEntity, CustomerEntity customerEntity, DealEntity dealEntity,
@@ -29,5 +40,5 @@ public abstract class SupplyMapper {
                 .weight(weight)
                 .date(dealDate)
                 .build();
-    };
+    }
 }