+++ src/main/java/com/example/demo/model/dto/response/SupplyReportResponse.java
@@ -1,5 +1,6 @@
 package com.example.demo.model.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,19 +8,22 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.util.List;
 
 @Data
 @SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class SupplyResponse {
+public class SupplyReportResponse {
 
-    private String name;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
+    private LocalDate date;
 
-    private BigDecimal pricePaid;
+    private double countWeight;
 
-    private double weight;
+    private BigDecimal countCost;
 
-    private BigDecimal cost;
+    private List<SupplyResponse> supplies;
 }